+++ backend/canisters/user_index/api/src/updates/mod.rs
@@ -1,5 +1,6 @@
 pub mod add_local_user_index_canister;
 pub mod add_platform_moderator;
+pub mod add_platform_operator;
 pub mod c2c_migrate_user_principal;
 pub mod c2c_notify_events;
 pub mod c2c_notify_low_balance;
@@ -12,6 +13,7 @@ pub mod mark_suspected_bot;
 pub mod pay_for_diamond_membership;
 pub mod register_user;
 pub mod remove_platform_moderator;
+pub mod remove_platform_operator;
 pub mod remove_sms_messages;
 pub mod set_governance_principals;
 pub mod set_max_concurrent_user_canister_upgrades;