+++ backend/canisters/user_index/impl/src/queries/mod.rs
@@ -4,6 +4,7 @@ pub mod check_username;
 pub mod current_user;
 pub mod http_request;
 pub mod platform_moderators;
+pub mod platform_operators;
 pub mod search;
 pub mod suspected_bots;
 pub mod user;