+++ backend/canisters/group_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Rename is_super_admin to is_platform_operator in c2c_lookup_user ([#3264](https://github.com/open-ic/open-chat/pull/3264)) 
+
 ### Removed
 
 - Removed code only needed for previous upgrade ([#3262](https://github.com/open-ic/open-chat/pull/3262))